+++ src/renderer/pages/replays/replay_file_stats/kill_table/kill_table.tsx
