+++ src/renderer/pages/replays/replay_file_stats/overall_table/overall_table.tsx
