+++ src/renderer/pages/replays/replay_file_stats/punish_table/punish_table.tsx
